requirements.txt
python-dotenv==1.0.1
google-generativeai==0.3.2
langchain>=0.0.350
langchain-core>=0.1.29,<0.2
langchain-community>=0.0.10
langchain-google-genai==0.0.6
pydantic>=2.0.0,<3.0.0
chainlit==1.3.2
streamlit>=1.31.0
protobuf>=4.25.3
google-api-core>=2.11.0
google-auth>=2.22.0
typing-extensions>=4.5.0
